+++ packages/core/src/connectors/xverse.ts
@@ -1,10 +1,12 @@
 import { SatsConnectConnector } from "~/connectors/sats-connect";
 
 export const createXverseConnector = () => {
+	const defaultProviderId = "XverseProviders.BitcoinProvider";
+
 	const providerId =
 		typeof window !== "undefined"
 			? window.btc_providers?.find((it) => it.name.includes("Xverse"))?.id
 			: "XverseProviders.BitcoinProvider";
 
-	return new SatsConnectConnector(providerId, "Xverse");
+	return new SatsConnectConnector(providerId || defaultProviderId, "Xverse");
 };