+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./constants";
 export * from "./createConfig";
 export * from "./networks";
 export * from "./utils";
+export * from "./providers";