+++ packages/react/src/hooks/useRunes.ts
@@ -1,13 +1,14 @@
 import {
+	type Config,
 	type GetRunesParams,
-	type GetRunesResponse,
+	type RunesResponse,
 	getRunes,
 } from "@midl-xyz/midl-js-core";
 import { type UseQueryOptions, useQuery } from "@tanstack/react-query";
-import { useMidlContext } from "~/context";
+import { useConfigInternal } from "~/hooks/useConfigInternal";
 
 type QueryOptions = Omit<
-	UseQueryOptions<GetRunesResponse>,
+	UseQueryOptions<RunesResponse>,
 	"queryFn" | "queryKey"
 > & {
 	queryKey?: ReadonlyArray<unknown>;
@@ -15,6 +16,7 @@ type QueryOptions = Omit<
 
 type UseRunesParams = GetRunesParams & {
 	query?: QueryOptions;
+	config?: Config;
 };
 
 /**
@@ -36,13 +38,18 @@ export const useRunes = ({
 	limit,
 	offset,
 	query: { queryKey, ...query } = {} as QueryOptions,
+	config: customConfig,
 }: UseRunesParams) => {
-	const { config } = useMidlContext();
+	const config = useConfigInternal(customConfig);
 
-	const { data: runes, ...rest } = useQuery<GetRunesResponse>({
+	const { data: runes, ...rest } = useQuery<RunesResponse>({
 		queryKey: ["runes", address, limit, offset, ...(queryKey ?? [])],
 		queryFn: () => {
-			return getRunes(config, { address, limit, offset });
+			return getRunes(config, {
+				address,
+				limit,
+				offset,
+			});
 		},
 		...query,
 	});