+++ packages/react/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./context";
 export * from "./hooks";
+export * from "./utils";
+export * from "./types";