+++ packages/react/src/utils/createStore.ts
@@ -1,0 +1,8 @@
+import { createStore as createZustandStore } from "zustand/vanilla";
+import type { MidlContextState } from "~/types";
+
+export const createStore = (initialState: MidlContextState) => {
+	return createZustandStore<MidlContextState>()(() => ({
+		...initialState,
+	}));
+};